+++ app/src/main/java/com/architectcoders/equipocinco/ui/MoviesFragment.kt
@@ -4,11 +4,20 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
+import androidx.navigation.NavController
+import androidx.navigation.findNavController
+import androidx.navigation.fragment.findNavController
+import androidx.recyclerview.widget.GridLayoutManager
+import com.architectcoders.data.Movie
 import com.architectcoders.equipocinco.R
+import com.architectcoders.generic.util.toast
 import com.architectcoders.presentation.di.modules.ViewModelProviderFactory
 import com.architectcoders.presentation.viewmodels.MovieViewModel
+import kotlinx.android.synthetic.main.fragment_movies.*
 import javax.inject.Inject
 
 /**
@@ -18,14 +27,44 @@ import javax.inject.Inject
 
 class MoviesFragment : Fragment() {
 
-    @Inject
-    lateinit var viewModelFactory: ViewModelProviderFactory
+    private lateinit var navController : NavController
 
     private val viewModel by lazy {
         ViewModelProvider(
             this,
-            viewModelFactory
+            (activity as MainActivity).viewModelFactory
         ).get(MovieViewModel::class.java)
+    }
+
+    private var adapter: MovieAdapter? = null
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        navController = view.findNavController()
+        viewModel.model.observe(this, Observer(::updateUI))
+    }
+
+    private fun updateUI(model: MovieViewModel.UiModel) {
+        when (model) {
+            is MovieViewModel.UiModel.RequestMovies -> viewModel.onRequestMovieList()
+            is MovieViewModel.UiModel.Content -> updateData(model.movies)
+        }
+    }
+
+    private fun updateData(movies: List<Movie>) {
+        initAdapter(movies)
+    }
+
+    private fun initAdapter(items: List<Movie>) {
+        rv?.let {
+            rv.layoutManager = GridLayoutManager(activity, 3)
+            adapter = MovieAdapter(items.toMutableList()) {
+                val action = R.id.action_moviesFragment_to_detailFragment
+                navController.navigate(action,
+                    bundleOf("movie" to it))
+            }
+            rv.adapter = adapter
+        }
     }
 
     override fun onCreateView(
