+++ app/src/main/res/layout/fragment_movies.xml
@@ -12,4 +12,5 @@
         tools:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
         tools:listitem="@layout/item_movie"
         tools:spanCount="3" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
