+++ buildsystem/dependencies.gradle
@@ -41,6 +41,7 @@ ext {
 
     navigationFragmentVersion = "2.1.0"
     navigationUiVersion = "2.1.0"
+    safeargs_version = "2.1.0-rc01"
 
     glideVersion = "4.9.0"
 
@@ -58,6 +59,7 @@ ext {
     classpathDependencies = [
             androidToolsBuild : "com.android.tools.build:gradle:${gradle_version}",
             kotlinGradlePlugin: "org.jetbrains.kotlin:kotlin-gradle-plugin:${kotlin_version}",
+            safeArgs: "androidx.navigation:navigation-safe-args-gradle-plugin:${safeargs_version}"
     ]
 
     androidXDependencies = [
